# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoUploadInput {
  /** Original video file in MP4, MOV, or WebM format. */
  video: Blob;
}

export interface SceneFrame {
  id: string;
  /** @minimum 1 */
  index: number;
  /** @minimum 0 */
  timestamp: number;
  timestampLabel: string;
  imageUrl: string;
  filename?: string;
}

export interface VideoBreakdown {
  id: string;
  filename: string;
  /** @minimum 0 */
  duration: number;
  /** @minimum 1 */
  width: number;
  /** @minimum 1 */
  height: number;
  originalVideoUrl: string;
  scenes: SceneFrame[];
}

export interface ErrorResponse {
  error: string;
}