# lib/api-zod/src/generated/types/sceneFrame.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SceneFrame {
  id: string;
  /** @minimum 1 */
  index: number;
  /** @minimum 0 */
  timestamp: number;
  timestampLabel: string;
  imageUrl: string;
  filename?: string;
}