# lib/api-zod/src/generated/types/videoBreakdown.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SceneFrame } from './sceneFrame';

export interface VideoBreakdown {
  id: string;
  filename: string;
  /** @minimum 0 */
  duration: number;
  /** @minimum 1 */
  width: number;
  /** @minimum 1 */
  height: number;
  originalVideoUrl: string;
  scenes: SceneFrame[];
}