# lib/api-zod/src/generated/types/videoUploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoUploadInput {
  /** Original video file in MP4, MOV, or WebM format. */
  video: Blob;
}